sync-goods-features/requirements.txt
elasticsearch~=7.14.0
SQLAlchemy~=1.4.22
pyodbc~=4.0.31
pandas~=1.1.4
numpy~=1.19.3
PyYAML~=5.4.1
